fix(regex-detector): Fix MemoryError caused by recrusive replace and                      limit max_matches to 50 by default
When replacing something like 'ab' -> 'abab' or '' with anything in a loop can grow string
causing more and more allocations, eventually crashing with MemoryError.
We fix this by doing careful replacements with a special placeholder and
also discarding empty matches and limiting matches to keep the detector
fast. The max_matches part is configurable via __init__
https://hello-haptik.atlassian.net/browse/ML-1443